% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleSetTests.R
\name{EnsembleSetTests}
\alias{EnsembleSetTests}
\title{Ensemble Set based tests for testing the association between a set of genetic variants and a phenotype}
\usage{
EnsembleSetTests(
  G,
  obj,
  B = 100,
  tests = c("Burden", "SKAT", "MORST"),
  weights.prior = "none",
  maf.beta.para = c(1, 25),
  is.pvals.path = FALSE,
  alpha = 1e-06,
  tau.type = "approx",
  target_power = 0.5,
  n.points = 50
)
}
\arguments{
\item{G}{a numeric matrix or dgCMatrix with each row as a different individual and each column as a separate gene/snp. Each genotype should be coded as 0, 1, 2.}

\item{obj}{an output from \code{\link{Null_model_glm}} or the function \code{\link{glmmkin}} from
the \code{\link{GMMAT}} package. See details.}

\item{B}{the number of base tests.}

\item{tests}{the names of set-based tests. It could only be one or several of “Burden”, “SKAT” and “MORST”.}

\item{weights.prior}{a character or a numeric vector. It is used to specify the standardized deviations (SDs) of normal variables when generating the random weights.
If it is character, it could only be \emph{“none”}, which means the SDs are all equal, or \emph{“beta”}, which means the maf-based beta weights will be used as the SDs and the parameters of beta weights are specified by \emph{maf.beta.para}.
If it is a numeric vector, it is the user-specified SDs and hence must have a length equal to ncol(G).}

\item{maf.beta.para}{a vector of parameters for the beta weights for the weighted kernels. It is only used when \emph{weights.prior == "beta"}.}

\item{is.pvals.path}{logical. If \emph{is.pvals.path == TRUE}, the p-value path of the ensemble test as \emph{B} increases will be provided in the output and can be used to draw the ensemble p-value path plot (See examples).}

\item{alpha}{the alpha parameter in MORST. It is suggested to be the significance level.}

\item{tau.type}{either "minimax" or "approx". See \code{\link{tau_c}} for details.}

\item{target_power}{a value that is used when \emph{tau.type == "approx"}. See \code{\link{tau_c}} for details.}

\item{n.points}{number of grid points used when \emph{tau.type == "minimax"}. See \code{\link{tau_c}} for details.}
}
\value{
A list with components:
\describe{
\item{\code{pval.ensemble.test}}{The p-value of the ensemble test}
\item{\code{pval.base.test}}{The p-values of the individual base tests}
\item{\code{pval.path.ensemble}}{The p-value path of the ensemble test when \emph{is.pvals.path == TRUE}.}
}
}
\description{
Calculate the ensemble Burden, ensemble SKAT, and ensemble MORST p-values under one or multiple sets of weights in GLMs or GLMMs.
}
\details{
If you want to fit a GLM, please use the \code{\link{Null_model_glm}} function to obtain the null model \emph{obj}.
If you have a kinship matrix/GRM and would like to fit a GLMM, please use the the function \code{\link{glmmkin}} from the the \code{\link{GMMAT}} package. Both dense and sparse kinship/GRM can be used.



While the \emph{alpha} parameter is suggested to be the significance level, practically there is no need to set \emph{alpha} less than 1e-08.
In most situations, the MORST p-values would only have negligible difference for values of \emph{alpha} less than 1e-06.
A super small \emph{alpha} could slow down the computation and might cause some numerical issue. Therefore, the default value for \emph{alpha} is 1e-06.
}
\examples{
 library(Matrix)
 data(Geno)
 G<-Geno[,1:100] # Geno is a dgCMatrix of genotypes
 Y<-rnorm(nrow(G)); Z<-matrix(rnorm(nrow(G)*4),ncol=4)
 obj<-Null_model_glm(Y,Z,family="gaussian")

 ### The number of base tests B = 10.
 EnsembleSetTests(G,obj,B=10)

 ### Use the beta prior, and only do Burden test
 EnsembleSetTests(G,obj,B=20,tests = "Burden",weights.prior = "beta")

 ### User-specified weights prior
 EnsembleSetTests(G,obj,B=20,tests = c("Burden","SKAT"),weights.prior = runif(ncol(G)))

 ### plot the p-value path of the ensemble tests as the number of base tests B increases
 set.seed(123); Y<- rowMeans(G)*4 + rnorm(nrow(G))
 Z<-matrix(rnorm(nrow(G)*4),ncol=4);obj<-Null_model_glm(Y,Z,family="gaussian")
 res <- EnsembleSetTests(G,obj,B=500,tests = c("Burden","SKAT"),is.pvals.path = TRUE,weights.prior = "beta")
 test.name = "Burden"; pvals.path <- res[["pval.path.ensemble"]][,test.name]
 plot(c(1:length(pvals.path)), -log10(pvals.path),xlab = "Number of base tests",ylab = "-log10(p-value)",type = "l")


}
\references{
Liu, Y., Liu, Z., and Lin, X. (2024+) Ensemble methods for testing a global null.
\emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}. Accepted.
}
\author{
Yaowu Liu
}
