% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Q_pval.R
\name{Get_Q_pval}
\alias{Get_Q_pval}
\title{Get the p-value of a quadratic test}
\usage{
Get_Q_pval(Q, w)
}
\arguments{
\item{Q}{The value of the quadratic test statistic. It must be positive.}

\item{w}{a numeric vector of positive weights. See details.}
}
\value{
The p-value of the quadratic test.
}
\description{
The survival function (i.e., p-value) of the weighted sum of i.i.d. Chi-squared variables with df=1 (i.e., a quadratic test).
A hybrid of davies's method, saddle point method, and liu's method is used.
}
\details{
Compute \eqn{P[T>Q]}, where \eqn{T=\sum_{i=1}^k w_iX_i}, \eqn{X_i}'s are i.i.d. Chi-squared variables with df=1, and \eqn{w_i}'s are the elements of \emph{w}.

A hybrid of several methods is used to acheive both accuracy and computation effeciency. Specifically, when the p-value > 1e-10, the davies' method is used;
when 1e-10 < p-value < 1e-15, the saddle point method is used; when p-value < 1e-15, liu's method is used.
}
\examples{
Get_Q_pval(50,seq(1,10,1))
Get_Q_pval(200,seq(1,10,1))
Get_Q_pval(1000,seq(1,10,1))
}
\references{
Davies, R. B. (1980). Algorithm AS 155: The distribution of a linear combination of X2 random variables.
\emph{Journal of the Royal Statistical Society. Series C (Applied Statistics)}, 29(3):323-333.

Kuonen, D. (1999). Miscellanea. saddlepoint approximations for distributions of quadratic forms in normal variables.
\emph{Biometrika}, 86(4):929-935.

Liu, H., Tang, Y., and Zhang, H. H. (2009). A new chi-square approximation to the distribution of
non-negative definite quadratic forms in non-central normal variables.
\emph{Computational Statistics & Data Analysis},53(4):853-856.
}
\author{
Yaowu Liu
}
