% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Null_model_glm.R
\name{Null_model_glm}
\alias{Null_model_glm}
\title{Fit the GLM null model for MORST}
\usage{
Null_model_glm(Y, Z, family = "gaussian", include_intercept = TRUE)
}
\arguments{
\item{Y}{a numeric vector of outcomes.}

\item{Z}{a numeric matrix of covariates that need to be adjusted.}

\item{include_intercept}{logical. If \env{TRUE}, the intercept will be included in the null model.}

\item{famlily}{a character. Should be "gaussian", "binomial" or "poisson". For each famliy, the cannonical link is used.}
}
\value{
This function returns an object that has model parameters and residuals of the NULL model of no association between outcomes Y and predictors X after adjusting for covariates Z. After obtaining it, please use \code{\link{MORST_glm}} or \code{\link{SetBasedTests}} to conduct the association test.
}
\description{
fit a GLM null model
}
\examples{
X<-matrix(rnorm(20000),ncol=20); Z=matrix(rnorm(nrow(X)*4),ncol=4)

### linear regression for continuous outcome
Y<-rnorm(nrow(X));obj<-Null_model_glm(Y,Z,family="gaussian")

### Logistic regression for binary outcome
Y<-rbinom(nrow(X),1,0.4);obj<-Null_model_glm(Y,Z,family="binomial")

### Binomial outcome
Y<-rbinom(nrow(X),5,0.4);Y<-cbind(Y,5-Y);obj<-Null_model_glm(Y,Z,family="binomial")

### Poisson outcome
Y<-rpois(nrow(X),5);obj<-Null_model_glm(Y,Z,family="poisson")
}
\author{
Yaowu Liu
}
