% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetBasedTests.R
\name{SetBasedTests}
\alias{SetBasedTests}
\title{Set based tests for testing the association between a set of genetic variants and a phenotype}
\usage{
SetBasedTests(
  G,
  obj,
  alpha = 1e-06,
  weights.beta = matrix(c(1, 25, 1, 1), nrow = 2),
  weights = NULL,
  tau.type = "approx",
  target_power = 0.5,
  n.points = 50,
  mac.thresh = 10
)
}
\arguments{
\item{G}{a numeric matrix or dgCMatrix with each row as a different individual and each column as a separate gene/snp. Each genotype should be coded as 0, 1, 2.}

\item{obj}{an output from \code{\link{Null_model_glm}} or the function \code{\link{glmmkin}} from
the \code{\link{GMMAT}} package. See details.}

\item{alpha}{the alpha parameter in MORST. It is suggested to be the significance level.}

\item{weights.beta}{a numeric vector/matrix of parameters for the beta weights for the weighted kernels. If it is a matrix, each column corresponds to one set of the beta-weights parameters.
If you want to use your own weights, please use the “weights” parameter. It will be ignored if “weights” parameter is not null.}

\item{weights}{a numeric vector/matrix of weights for the SNPs. If it is a matrix, each column corresponds to one set of weights. When it is NULL, the beta weight with the “weights.beta” parameter is used.}

\item{tau.type}{either "minimax" or "approx". See \code{\link{tau_c}} for details.}

\item{target_power}{a value that is used when \emph{tau.type == "approx"}. See \code{\link{tau_c}} for details.}

\item{n.points}{number of grid points used when \emph{tau.type == "minimax"}. See \code{\link{tau_c}} for details.}

\item{mac.thresh}{a threshold of the minor allele count (MAC) that is used in the ACAT-V test. SNPs with MAC less than this thrshold will be frist aggregated by the Burden test in ACAT-V.}
}
\value{
The p-values of Burden, SKAT, ACAT-V and MORST under under one or multiple choices of weights.
}
\description{
Calculate the Burden, SKAT, ACAT-V and MORST p-values under one or multiple sets of weights in GLMs or GLMMs.
}
\details{
If you want to fit a GLM, please use the \code{\link{Null_model_glm}} function to obtain the null model \emph{obj}.
If you have a kinship matrix/GRM and would like to fit a GLMM, please use the the function \code{\link{glmmkin}} from the the \code{\link{GMMAT}} package. Both dense and sparse kinship/GRM can be used.

The ACAT-V p-value might be slightly different from the result from the \code{\link{ACAT_V}} function in the \code{\link{ACAT}} package. This is because the variant-level p-values are calculated using slighly different methods.

While the \emph{alpha} parameter is suggested to be the significance level, practically there is no need to set \emph{alpha} less than 1e-08.
In most situations, the MORST p-values would only have negligible difference for values of \emph{alpha} less than 1e-06.
A super small \emph{alpha} could slow down the computation and might cause some numerical issue. Therefore, the default value for \emph{alpha} is 1e-06.
}
\examples{
 library(Matrix)
 data(Geno)
 G<-Geno[,1:100] # Geno is a dgCMatrix of genotypes
 Y<-rnorm(nrow(G)); Z<-matrix(rnorm(nrow(G)*4),ncol=4)
 obj<-Null_model_glm(Y,Z,family="gaussian")
 SetBasedTests(G,obj)
}
\author{
Yaowu Liu
}
