% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_c.R
\name{tau_c}
\alias{tau_c}
\title{The choice of tau in MORST}
\usage{
tau_c(eg.values, alpha, tau.type = "approx", target_power = 0.5, n.points = 50)
}
\arguments{
\item{eg.values}{a numeric vector of non-negative eigenvalues.}

\item{alpha}{the alpha parameter in MORST. It is suggested to be the significance level.}

\item{tau.type}{either "minimax" or "approx". If \emph{tau.type == "minimax"}, tau is calculated based on the minimax criterion;
if \emph{tau.type == "approx"}, tau is an approximation to the minimax tau. Default value is "approx".}

\item{target_power}{a value that is used when \emph{tau.type == "approx"}. See details.}

\item{n.points}{number of grid points used when \emph{tau.type == "minimax"}. Should be at least 20 to have reasonable accuracy.}
}
\value{
the parameter tau in MORST
}
\description{
Calculate the parameter tau in MORST based on either the minimax criterion or an approximation of the minimax solution
}
\details{
The approximation method is substantially faster than the minimax method and will be introduced soon.
}
\examples{
tau_c(seq(1,100,1),1e-04)
tau_c(seq(1,100,1),1e-04,tau.type = "minimax")
tau_c(c(2,rep(0.2,20)),0.05,tau.type = "minimax",n.points = 200)
}
\references{
Liu, Y., Li, Z., and Lin, X. (2020+) A Minimax Optimal Ridge-Type Set Test
for Global Hypothesis with Applications in Whole Genome Sequencing Association Studies.
\emph{Journal of the American Statistical Association}. Accepted.
}
\author{
Yaowu Liu
}
